Code fixes.
Fixed the testing for plot paths, which should use all() so that all criteria are fulfilled, before plotting is skipped.
Fixed the naming of dispersion and MA plots for genes.
Fixed the exclusion of the first, redundant identifier column in data.frame objects retrieved from cummeRbund::diffData(). This should be fixed up-stream really.