Drop the log() function.
Filter NA values from the gene set tibble for custom annotation.
Drop the log() of the DESEq2 variance stabilising transformed count matrix, since it is already on a log2 scale.
Replace NaN values in the matrix after scaling, caused by rows with identical values, with 0.0 values.
Reverted the selection of gene_name values for the top gene identifiers back to data frame syntax as it is based on row names.